Form code of plugins optimized
